+++ pytorch_lightning/trainer/auto_mix_precision.py
@@ -20,12 +20,6 @@ class TrainerAMPMixin(ABC):
     use_native_amp: bool
 
     def init_amp(self, use_amp):
-        # TODO: remove in v 0.8.0
-        if self.use_native_amp:
-            rank_zero_warn("`amp_level` has been deprecated since v0.7.4 "
-                           "(native amp does not require it)"
-                           " and this argument will be removed in v0.8.0", DeprecationWarning)
-
         # Backward compatibility, TODO: remove in v0.9.0
         if use_amp is not None:
             rank_zero_warn("`use_amp` has been replaced by `precision` since v0.7.0"
@@ -37,16 +31,6 @@ class TrainerAMPMixin(ABC):
         if use_amp and self.use_native_amp:
             log.info('Using 16bit precision.')
             return
-
-        # TODO: remove all below for v0.8.0
-        if use_amp and not APEX_AVAILABLE:  # pragma: no-cover
-            raise ModuleNotFoundError("""
-            You set `use_amp=True` but do not have apex installed.
-            Install apex first using this guide and rerun with use_amp=True:
-            https://github.com/NVIDIA/apex#linux
-
-            this run will NOT use 16 bit precision
-            """)
 
         if self.use_amp:
             log.info('Using 16bit precision.')